third_party/blink/web_tests/fast/events/focus-event-source-device-from-touch-expected.txt

Test that focus events from a touch device should have a sourceCapabilities and its firesTouchEvents should be true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

tests on div
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true


tests on input
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true


tests on button
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true


tests on textarea
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true


tests on a
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true


PASS successfullyParsed is true

TEST COMPLETE
